+++ apps/evm/src/clients/api/__mocks__/index.ts
@@ -272,6 +272,11 @@ export const useGetLiquidityHubs = vi.fn(() => ({
   },
 }));
 
+export const useGetChainIdsWithIsolatedPoolPosition = vi.fn(() => ({
+  isLoading: false,
+  chainIds: [],
+}));
+
 export const getLiquidityHub = vi.fn(async () => ({
   liquidityHub: liquidityHubs[0],
 }));