+++ apps/evm/src/pages/IsolatedPools/testIds.ts
@@ -1,5 +1,0 @@
-const TEST_IDS = {
-  marketTable: 'dashboard-market-table',
-};
-
-export default TEST_IDS;