+++ packages/chains/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @venusprotocol/chains
 
+## 0.31.1
+
+### Patch Changes
+
+- d22d3c1: add support for vceBTC and Solv venue
+
+## 0.31.0
+
+### Minor Changes
+
+- e49100e: feat: update hBnb address
+
 ## 0.30.0
 
 ### Minor Changes